+++ src/Sulu/Bundle/MediaBundle/Api/Media.php
@@ -220,6 +220,14 @@ class Media extends ApiWrapper
     }
 
     /**
+     * @return string|null
+     */
+    public function getExtension()
+    {
+        return $this->getFileVersion()->getExtension();
+    }
+
+    /**
      * @param string $title
      *
      * @return $this