+++ CHANGELOG.md
@@ -1,10 +1,16 @@
 CHANGELOG for Sulu
 ==================
 
-* dev-master
+* 1.6.14 (2018-02-06)
     * ENHANCEMENT #3717 [ContentBundle]           ResourceLocator: Show whole url in history overlay
+    * HOTFIX      #3739 [ContentBundle]           Added locale to content-teaser query
+    * ENHANCEMENT #3735 [DocumentManager]         Set proper default locale for document-manager
+    * ENHANCEMENT #3736 [WebsiteBundle]           Added exception when default_host is needed for sitemap generation
+    * HOTFIX      #3741 [ContentBundle]           Cleand-up xliff import/export component
+    * HOTFIX      #3730 [ContactBundle]           Fixed class parameter to load field-descriptor
+    * HOTFIX      #3720 [MediaBundle]             Added extension-guesser to fix wrong extensions on download
 
-* 1.6.13 (2017-01-18)
+* 1.6.13 (2018-01-18)
     * ENHANCEMENT #3715 [RouteBundle]             RouteRepository: Added method 'findAllByEntity'
     * HOTFIX      #3713 [WebsiteBundle]           Fixed redirect listener to allow trailing slashes for homepage
     * HOTFIX      #3709 [CategoryBundle]          Category API: Fix bug when searching with rootKey parameter
@@ -154,6 +160,14 @@ CHANGELOG for Sulu
     * FEATURE     #3278 [Util]                  Changed way to determine `sulu.version` to composer.lock
     * FEATURE     #3278 [Util]                  Introduced new parameter `app.version` read from composer.json
     * ENHANCEMENT #3154 [All]                   Upgrade symfony to ^3.0
+
+* 1.5.10 (2018-02-06)
+    * HOTFIX      #3739 [ContentBundle]         Added locale to content-teaser query
+    * ENHANCEMENT #3735 [DocumentManager]       Set proper default locale for document-manager
+    * ENHANCEMENT #3736 [WebsiteBundle]         Added exception when default_host is needed for sitemap generation
+    * HOTFIX      #3741 [ContentBundle]         Cleand-up xliff import/export component
+    * HOTFIX      #3730 [ContactBundle]         Fixed class parameter to load field-descriptor
+    * HOTFIX      #3720 [MediaBundle]           Added extension-guesser to fix wrong extensions on download
 
 * 1.5.9 (2018-01-18)
     * HOTFIX      #3709 [CategoryBundle]        Category API: Fix bug when searching with rootKey parameter